Extract useDepositOf
