+++ packages/next-common/components/publicProposal/metadata.js
@@ -1,3 +1,4 @@
+import React from "react";
 import styled from "styled-components";
 import User from "../user";
 import Links from "../links";
